requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
future==0.18.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
PyVimeo==1.1.0
requests==2.26.0
six==1.16.0
tinydb==4.5.2
tuspy==0.2.5
urllib3==1.26.7
Werkzeug==2.0.2
